# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ursainyk API",
    "version": "0.1.0",
    "description": "Nabhahita Phase 1 — candidates, ESM centres, contractors, review gate, matching, verification, billing. Authorization is role-based (see x-required-permission per operation); scoping (territory/org/own) is enforced server-side."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request candidate OTP (auto-registers unknown phones)",
        "operationId": "post_auth_otp_request",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+[1-9]\\d{7,14}$"
                  }
                },
                "required": [
                  "phone"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify OTP → token pair",
        "operationId": "post_auth_otp_verify",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+[1-9]\\d{7,14}$"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "phone",
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Portal login (TOTP required for enrolled admins)",
        "operationId": "post_auth_login",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 128
                  },
                  "totp": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Rotate refresh token (reuse revokes the family)",
        "operationId": "post_auth_refresh",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 512
                  }
                },
                "required": [
                  "refreshToken"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke all refresh tokens",
        "operationId": "post_auth_logout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Current principal (roles, territories, orgs)",
        "operationId": "get_auth_me",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        }
      }
    },
    "/auth/password/change": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change own password (revokes all sessions)",
        "operationId": "post_auth_password_change",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 128
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 128
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/totp/enroll": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Start TOTP enrollment → otpauth URI",
        "operationId": "post_auth_totp_enroll",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        }
      }
    },
    "/auth/totp/activate": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Prove TOTP possession — enforced at next login",
        "operationId": "post_auth_totp_activate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "code"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/password/set": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Redeem single-use set-password link (invite/reset)",
        "operationId": "post_auth_password_set",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed (zod issues in body)"
          },
          "401": {
            "description": "Missing/invalid token, or TOTP required"
          },
          "403": {
            "description": "Role lacks the required permission"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 200
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 128
                  }
                },
                "required": [
                  "token",
                  "newPassword"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  }
}